requirements.txt
numpy
tensorflow-gpu
tensorflow-graphics-gpu
Pillow
rawpy
matplotlib.pyplot
